editoast: move `views::Server` in its own submodule
